requirements.txt
google-adk
google-generativeai
python-dotenv
fastapi
uvicorn
python-magic
nest_asyncio
